Add support for config tree values in templates.
You can now register a list of names and paths for a view. Each path
should correspond with a config tree path. These config values will then
be available in the template for that view under the name you set for
it.

This has first been implemented so as to put the name of the site into
the navbar.