Expose the pools in use, and the cache object, from the config object.
The primary use of the result from doing the latter is that helpers will
be able to access the cache. Ideally there would be a way to pass
helpers the cache object directly, but I haven't quite worked out a way
to do this in a good way yet. I imagine the answer will become clearer
once PHP 5.6 comes out and I can use variadic functions and argument
unpacking.