Added ability to override pretty much all config settings.
Previously, if a setting was in a group of numerically-indexed items, it
wasn't possible to override the settings inside them.

Now, if you have a config node in the same place, and the "name"
attributes are the same, the values in the later area will override the
values in the older area.