Added Aura DI module.
Currently, the DI container is used solely for controllers to retrieve
instances of objects it requires to use while rendering a page (or
whatever else it chooses to do). I may end up redoing the helper
system at some point to use the DI container.

I'm not 100% sure that I'm using the DI container correctly, but I
really don't see the point of creating a whole separate object to act as
a factory when the DI container can do that for me just fine. I just
wish there was a cleaner way to do it than calling __invoke() on the
returned object manually.