Created a generic way of adding meta tags to pages.
This addition involves a few things.

Firstly, the routes.json config file is back. This time, it actually
does something! At the moment, it's used solely to define default meta
tag values for individual routes.

Secondly, I had to update the mattyg/meta package so that it actually
worked. Lol.

Thirdly, I added default meta tag values to the site configuration.

Fourth, I added a function to the core abstract handler to get the
default values for meta tags for a route. It gets the defaults for the
sites, then overwrites any necessary tags with ones for the current
route. It does not automatically set these values on the meta helper, so
that subclasses can override the function for their own purposes.